feat(update): more layout updates
