+++ src/app/(pages)/kb/[section]/article/[slug]/page.tsx
@@ -1,4 +1,4 @@
-import { ArticleLayout } from '../../../layouts/ArticleLayout';
+import { ArticleLayout } from 'components/Layouts/Article';
 
 export default async function KnowledgeBase() {
     return <ArticleLayout />
