+++ src/components/Layouts/Knowledgebase/index.tsx
@@ -1,21 +1,19 @@
 'use client';
 
 import Link from 'next/link';
-import { motion } from "framer-motion";
 import { FC } from 'react';
+import { motion } from "framer-motion";
 import { MdArticle } from 'react-icons/md';
-import { usePathname } from 'next/navigation';
-import { PageHero } from '@/src/app/components/PageHero/UsePageHero';
+import { PageHero } from 'components/PageHero';
 import { githubFetcher } from '@/src/fetchers/github';
-import { useSWRClient } from '@/providers/SWR/config';
-import ErrorLayout from '@/src/app/components/Static/ErrorLayout';
-import LoadingSkeleton from './LoadingSkeleton';
+import { useSWRClient } from 'providers/SWR/config';
+import ErrorLayout from 'components/Static/ErrorLayout';
+import LoadingSkeleton from '../LoadingSkeleton';
 
 interface Article {
     slug: string;
     title: string;
     description: string;
-    section: string;
     path: string;
 }
 
@@ -26,42 +24,24 @@ interface Section {
     articles?: Article[];
 }
 
-export const ArticleSectionLayout: FC = ({ }) => {
-    const slug = usePathname().replace('/kb/', '');
-
+export const KnowledgeBaseLayout: FC = ({ }) => {
     const { data, error, isLoading } = useSWRClient(
         {
             repoOwner: 'NodeByteHosting',
             repoName: 'assets/contents',
             jsonPath: 'markdown/kb/articles.json?ref=main',
-            slug,
-            type: 'section',
+            type: 'articles',
         },
         githubFetcher
     );
 
-    const section: Section = data || {};
-    const articles: Article[] = section.articles || [];
-
-    const pageTitle = error
-        ? 'Internal Error'
-        : isLoading
-            ? 'Loading Articles...'
-            : section.section
-            || 'Loading...';
-
-    const pageSubtitle = error
-        ? `An error occurred: ${error.message}`
-        : isLoading
-            ? 'Please wait while we fetch the articles.'
-            : section.about
-            || 'Please wait while we fetch the articles.';
+    const sections: Section[] = data || [];
 
     return (
         <>
             <PageHero
-                title={pageTitle}
-                text={pageSubtitle}
+                title="Knowledge Base"
+                text="Find all the information you need to use our services, from getting started to advanced configurations."
             />
             <motion.section className="py-16 bg-dark">
                 <div className="px-4 mx-auto max-w-screen-xl">
@@ -69,25 +49,27 @@ export const ArticleSectionLayout: FC = ({ }) => {
                         <ErrorLayout />
                     ) : isLoading || !data ? (
                         <LoadingSkeleton />
-                    ) : articles.length === 0 ? (
+                    ) : sections.length === 0 ? (
                         <div className="text-center text-white">No articles available at the moment.</div>
                     ) : (
                         <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 gap-8">
-                            {articles.map((article, index) => (
-                                <Link key={index} href={`/kb/${article.section}/article/${article.slug}`}>
+                            {sections.map((section, index) => (
+                                <Link key={index} href={`/kb/${section.slug}`}>
                                     <article className="p-6 relative shadow bg-dark_gray hover:bg-black_secondary rounded-lg border border-gray-200 dark:bg-gray-800 dark:border-gray-700 flex flex-col justify-between">
                                         <div>
                                             <div className="flex justify-between items-center mb-4 text-gray-500">
                                                 <span className="text-sm flex items-center text-white/50">
                                                     <MdArticle className="mr-1 text-white/50" />
-                                                    Article
+                                                    {section.articles?.length} Articles
                                                 </span>
                                             </div>
                                             <h2 className="mb-4 text-2xl font-bold tracking-tight text-white">
-                                                {article.title}
+                                                <Link href={`/kb/${section.slug}`}>
+                                                    {section.section}
+                                                </Link>
                                             </h2>
                                             <p className="mb-6 font-light text-white/50">
-                                                {article.description}
+                                                {section.about}
                                             </p>
                                         </div>
                                     </article>
