+++ src/components/Static/ErrorLayout.tsx
@@ -4,7 +4,7 @@ import Link from "next/link";
 import { motion } from "framer-motion";
 import { FiExternalLink } from "react-icons/fi";
 import { ERROR_PAGE_LINKS } from "../../utils/errorLinks";
-import { DynamicIcon } from "@/src/providers/DynamicIcon/DynamicIcon";
+import { DynamicIcon } from "@/src/providers/DynamicIcon";
 
 interface DynamicIconProps {
     iconName: string;
