+++ src/providers/DynamicIcon/index.tsx
