+++ src/providers/Session/index.tsx
@@ -1,0 +1,9 @@
+'use client';
+
+import { SessionProvider } from 'next-auth/react';
+
+const NodeByteSession = ({ children }: { children: React.ReactNode }) => {
+    return <SessionProvider>{children}</SessionProvider>;
+};
+
+export default NodeByteSession;
